+++ src/components/index.js
@@ -1,0 +1,3 @@
+export { Navbar } from './NavBar/Navbar';
+export { Footer } from './Footer/Footer';
+export { RequiresAuth } from './RequiresAuth';
