+++ src/index.js
@@ -4,11 +4,7 @@ import "./index.css";
 import App from "./App";
 import { makeServer } from "./server";
 import { BrowserRouter } from "react-router-dom";
-import { ProductContextProvider } from "./context/Product-context";
-import { WishlistProvider } from "./context/wishlist-context";
-import { CartListProvider } from "./context/cart-context";
-import { UserContextProvider } from "./context/userAuth-context";
-import { ToastContextProvider } from "./context/toastContext";
+import { ProductContextProvider, ToastContextProvider, UserContextProvider, CartListProvider, WishlistProvider } from "./context/index";
 
 
 // Call make Server
