+++ src/pages/index.js
@@ -1,0 +1,7 @@
+export { Cart } from './Cart/Cart';
+export { Home } from './HomePage/Home';
+export { ProductDetail } from './ProductDetailPage/ProductDetail';
+export { ProductListing } from './ProductListingPage/ProductListing';
+export { SignIn } from './SignIn/SignIn';
+export { SignUp } from './SignUp/SignUp';
+export { WishList} from './WishList/WishList';
